[Decomp][Player] Cohesion split: extract loot system to PlayerLoot.cpp
